feat: Add structured portfolio data and dedicated pages for projects, certifications, and publications.
