+++ app/projects/[slug]/page.tsx
@@ -22,10 +22,6 @@ import Link from 'next/link';
 import ProjectQR from '@/app/projects/components/project-qr';
 import ShareButtons from '@/app/projects/components/share-buttons';
 import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { oneDark } from 'react-syntax-highlighter/dist/cjs/styles/prism';
 import GalleryLightbox from '@/app/projects/components/gallery-lightbox';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import type { Metadata } from 'next';
