+++ lib/data/certifications.ts
@@ -1,0 +1,30 @@
+import { Certification } from './types';
+
+export const certifications: Certification[] = [
+  {
+    title: 'PCEP™ – Certified Entry-Level Python Programmer',
+    issuer: 'Python Institute',
+    date: 'Issued Jun 2024',
+    credentialId: 'PCEP-30-02',
+    link: 'https://www.credly.com/badges/c97d5448-24e6-4f37-80c1-b83ab768bbdd/linked_in_profile',
+    skills: ['Python Programming', 'Algorithm Design', 'Problem Solving'],
+    highlight: 'Mastered core Python concepts and best practices',
+  },
+  {
+    title: 'React Development Certification',
+    issuer: 'Codecademy',
+    date: 'Issued Feb 2025',
+    link: 'https://www.codecademy.com/profiles/Ad-Archer/certificates/af00e5032d0a68cc84879983f5d8333b',
+    skills: ['React.js', 'State Management', 'Component Architecture'],
+    highlight: 'Built 5+ production-ready React applications',
+  },
+  {
+    title: 'AI & Machine Learning Fundamentals',
+    issuer: 'Databricks',
+    date: 'Issued Nov 2024 · Expires Nov 2026',
+    credentialId: '121496255',
+    link: 'https://credentials.databricks.com/03505993-f39c-4a0e-9b60-d63684c156b6',
+    skills: ['AI Development', 'ML Models', 'Data Analysis'],
+    highlight: 'Applied AI concepts in real-world projects',
+  },
+];
